+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ## New features
 
+* Added `process_jags_model()` for modeling antibody response curves and assessing model fit (#68)
 * Diagnostic function to produce density plots with stratification (#27)
 * Added SEES data set data folder and documentation (#41)
 * Fixing SEES data and added jags_post for SEES (#63)